backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
resend==2.0.0
